Corrected casting

+++ include/project4/stat_utils.hpp
@@ -4,12 +4,12 @@
 #include <functional>
 
 namespace stat_utils{
-    void print_map(std::map<int, double> p);
-    double expected_value(std::map<int, double> p);
-    double expected_value(std::map<int, double> p, std::function<double(int)> f);
+    void print_map(std::map<double, double> p);
+    double expected_value(std::map<double, double> p);
+    double expected_value(std::map<double, double> p, std::function<double(double)> f);
     double expected_value(int* samples, int N);
     double expected_value(int* samples, int N, std::function<double(int)> f);
-    std::map<int, double> distribution(int* samples, int N);
-    std::map<int, double> distribution(int* samples, int N, std::function<double(int)> f);
+    std::map<double, double> distribution(int* samples, int N);
+    std::map<double, double> distribution(int* samples, int N, std::function<double(int)> f);
 }
