+++ src/utils.cpp
@@ -2,16 +2,16 @@
 #include <map>
 #include "project4/stat_utils.hpp"
 
-double I(int x) {return x;}
+double I(float x){return x;}
 
-void stat_utils::print_map(std::map<int, double> p){
+void stat_utils::print_map(std::map<double, double> p){
     for (auto const& bucket : p){
         std::cout << bucket.first << ": " << bucket.second << std::endl;
     }
 }
 
 
-double stat_utils::expected_value(std::map<int, double> p, std::function<double(int)> f){
+double stat_utils::expected_value(std::map<double, double> p, std::function<double(double)> f){
     double expected_value = 0.;
     for (auto const& bucket : p){
         expected_value += f(bucket.first) * bucket.second;
@@ -19,7 +19,7 @@ double stat_utils::expected_value(std::map<int, double> p, std::function<double(
     return expected_value;
 }
 
-double stat_utils::expected_value(std::map<int, double> p){
+double stat_utils::expected_value(std::map<double, double> p){
     return expected_value(p, I);
 }
 
@@ -35,8 +35,8 @@ double stat_utils::expected_value(int *samples, int N){
     return expected_value(samples, N, I);
 }
 
-std::map<int, double> stat_utils::distribution(int *samples, int N, std::function<double(int)> f){
-    std::map<int, double> buckets;
+std::map<double, double> stat_utils::distribution(int *samples, int N, std::function<double(int)> f){
+    std::map<double, double> buckets;
     for (int i = 0; i < N; i++){
         float sample = f(samples[i]);
         buckets.emplace(sample, 0);
@@ -45,6 +45,6 @@ std::map<int, double> stat_utils::distribution(int *samples, int N, std::functio
     return buckets;
 }
 
-std::map<int, double> stat_utils::distribution(int *samples, int N){
+std::map<double, double> stat_utils::distribution(int *samples, int N){
     return distribution(samples, N, I);
 }
